Move microphone recording out of TileService
